backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
supabase>=2.0.0
firebase-admin>=6.0.0
httpx>=0.25.0
python-multipart>=0.0.6
openai>=1.6.0
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=10.0.0

